Fixes MT#6624, as per the report
I've done a quick change to 6800ops.hxx that seems to fix the problem.  That said, this is a change that simply takes the report at its word, and I've made no attempt to validate the veracity of the report.  Furthermore, this code affects a number of CPU cores in MAME (mc6800, m6801 etc).  Before this gets merged (or even before I am willing to submit a PR), someone needs to do their homework to determine if this is a correct fix.